+++ datafusion/physical-expr/src/array_expressions.rs
@@ -229,120 +229,6 @@ fn check_datatypes(name: &str, args: &[&ArrayRef]) -> Result<()> {
     Ok(())
 }
 
-/// Convert one or more [`ArrayRef`] of the same type into a
-/// `ListArray` or 'LargeListArray' depending on the offset size.
-///
-/// # Example (non nested)
-///
-/// Calling `array(col1, col2)` where col1 and col2 are non nested
-/// would return a single new `ListArray`, where each row was a list
-/// of 2 elements:
-///
-/// ```text
-/// ┌─────────┐   ┌─────────┐           ┌──────────────┐
-/// │ ┌─────┐ │   │ ┌─────┐ │           │ ┌──────────┐ │
-/// │ │  A  │ │   │ │  X  │ │           │ │  [A, X]  │ │
-/// │ ├─────┤ │   │ ├─────┤ │           │ ├──────────┤ │
-/// │ │NULL │ │   │ │  Y  │ │──────────▶│ │[NULL, Y] │ │
-/// │ ├─────┤ │   │ ├─────┤ │           │ ├──────────┤ │
-/// │ │  C  │ │   │ │  Z  │ │           │ │  [C, Z]  │ │
-/// │ └─────┘ │   │ └─────┘ │           │ └──────────┘ │
-/// └─────────┘   └─────────┘           └──────────────┘
-///   col1           col2                    output
-/// ```
-///
-/// # Example (nested)
-///
-/// Calling `array(col1, col2)` where col1 and col2 are lists
-/// would return a single new `ListArray`, where each row was a list
-/// of the corresponding elements of col1 and col2.
-///
-/// ``` text
-/// ┌──────────────┐   ┌──────────────┐        ┌─────────────────────────────┐
-/// │ ┌──────────┐ │   │ ┌──────────┐ │        │ ┌────────────────────────┐  │
-/// │ │  [A, X]  │ │   │ │    []    │ │        │ │    [[A, X], []]        │  │
-/// │ ├──────────┤ │   │ ├──────────┤ │        │ ├────────────────────────┤  │
-/// │ │[NULL, Y] │ │   │ │[Q, R, S] │ │───────▶│ │ [[NULL, Y], [Q, R, S]] │  │
-/// │ ├──────────┤ │   │ ├──────────┤ │        │ ├────────────────────────│  │
-/// │ │  [C, Z]  │ │   │ │   NULL   │ │        │ │    [[C, Z], NULL]      │  │
-/// │ └──────────┘ │   │ └──────────┘ │        │ └────────────────────────┘  │
-/// └──────────────┘   └──────────────┘        └─────────────────────────────┘
-///      col1               col2                         output
-/// ```
-fn array_array<O: OffsetSizeTrait>(
-    args: &[ArrayRef],
-    data_type: DataType,
-) -> Result<ArrayRef> {
-    // do not accept 0 arguments.
-    if args.is_empty() {
-        return plan_err!("Array requires at least one argument");
-    }
-
-    let mut data = vec![];
-    let mut total_len = 0;
-    for arg in args {
-        let arg_data = if arg.as_any().is::<NullArray>() {
-            ArrayData::new_empty(&data_type)
-        } else {
-            arg.to_data()
-        };
-        total_len += arg_data.len();
-        data.push(arg_data);
-    }
-
-    let mut offsets: Vec<O> = Vec::with_capacity(total_len);
-    offsets.push(O::usize_as(0));
-
-    let capacity = Capacities::Array(total_len);
-    let data_ref = data.iter().collect::<Vec<_>>();
-    let mut mutable = MutableArrayData::with_capacities(data_ref, true, capacity);
-
-    let num_rows = args[0].len();
-    for row_idx in 0..num_rows {
-        for (arr_idx, arg) in args.iter().enumerate() {
-            if !arg.as_any().is::<NullArray>()
-                && !arg.is_null(row_idx)
-                && arg.is_valid(row_idx)
-            {
-                mutable.extend(arr_idx, row_idx, row_idx + 1);
-            } else {
-                mutable.extend_nulls(1);
-            }
-        }
-        offsets.push(O::usize_as(mutable.len()));
-    }
-    let data = mutable.freeze();
-
-    Ok(Arc::new(GenericListArray::<O>::try_new(
-        Arc::new(Field::new("item", data_type, true)),
-        OffsetBuffer::new(offsets.into()),
-        arrow_array::make_array(data),
-        None,
-    )?))
-}
-
-/// `make_array` SQL function
-pub fn make_array(arrays: &[ArrayRef]) -> Result<ArrayRef> {
-    let mut data_type = DataType::Null;
-    for arg in arrays {
-        let arg_data_type = arg.data_type();
-        if !arg_data_type.equals_datatype(&DataType::Null) {
-            data_type = arg_data_type.clone();
-            break;
-        }
-    }
-
-    match data_type {
-        // Either an empty array or all nulls:
-        DataType::Null => {
-            let array =
-                new_null_array(&DataType::Null, arrays.iter().map(|a| a.len()).sum());
-            Ok(Arc::new(array_into_list_array(array)))
-        }
-        DataType::LargeList(..) => array_array::<i64>(arrays, data_type),
-        _ => array_array::<i32>(arrays, data_type),
-    }
-}
 
 fn general_array_element<O: OffsetSizeTrait>(
     array: &GenericListArray<O>,
